% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massageData.r
\name{AddFactor}
\alias{AddFactor}
\title{Add a Factor}
\usage{
AddFactor(factorReturns, newFactorReturns, calcExcess = FALSE)
}
\arguments{
\item{factorReturns}{Tibble of existing returns with date in first column.}

\item{newFactorReturns}{Vector of new returns with date in first column.}

\item{calcExcess}{If FALSE, no adjustments are made. If TRUE, the newFactorReturns are subtracted from
each column of the factorReturns creating excess returns.}
}
\value{
Tibble with new column
}
\description{
Adds a column of returns to an existing table.  Optionally subtracts the returns from each column of the
existing table.
}
\details{
If the date columns do not align, the returning table will have the maximum common date range.  The 
first column of both tables of returns be date.
}
\examples{
AddFactor(factorReturns=factorReturns.US, newFactorReturns=newFactorReturns, calcExcess=TRUE)
}
