% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riingo.r
\name{convert_prices_to_monthly_returns}
\alias{convert_prices_to_monthly_returns}
\title{Convert Daily Prices to Monthly Returns}
\usage{
convert_prices_to_monthly_returns(prices)
}
\arguments{
\item{prices}{Vector of monthly prices}
}
\value{
Vector of monthly returns. First element will be NA.
}
\description{
Convert Daily Prices to Monthly Returns
}
\examples{
convert_prices_to_monthly_returns(rnorm(30, 100, 1))
}
