% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{factorGroups}
\alias{factorGroups}
\title{Factor groups}
\format{A list with 3 items:
\describe{
  \item{US}{US factor names. }
  \item{Intl}{International developed factor names}
  \item{EM}{Emerging market factor names}
}}
\usage{
factorGroups
}
\description{
A 3 item list grouping the factorReturns variables into regions.
}
\keyword{datasets}
