% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{factorReturns.EM}
\alias{factorReturns.EM}
\title{RAFI Emerging Market Factor Returns}
\format{A tibble with 381 rows and 15 variables:
\describe{
  \item{date}{date at the end of the month}
  \item{EMVal}{Emerging Value factor}
  \item{EMLVol}{Emerging Low volatility factor}
  \item{EMQual}{Emerging Quality factor}
  \item{EMMom}{Emerging Momentum factor}
}}
\usage{
factorReturns.EM
}
\description{
A dataset containing the monthly returns of 4 factors across emerging markets.
}
\keyword{datasets}
