% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massageData.r
\name{FindMaxCommonDateRange}
\alias{FindMaxCommonDateRange}
\title{Find Maximum Common Date Range
Finds the maximum common date range between two vectors of dates}
\usage{
FindMaxCommonDateRange(dates1, dates2)
}
\arguments{
\item{dates1}{Vector of dates}

\item{dates2}{Vector of dates}
}
\value{
Vector of two dates: start and end
}
\description{
Find Maximum Common Date Range
Finds the maximum common date range between two vectors of dates
}
\examples{
FindMaxCommonDateRange(factorReturns$date, factorReturns$date)
}
