% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.r
\name{get_lower_tri}
\alias{get_lower_tri}
\title{Get lower triangle of a correlation matrix}
\usage{
get_lower_tri(cormat)
}
\arguments{
\item{cormat}{Correlation matrix}
}
\value{
Lower triangle of matrix, upper is NA
}
\description{
Get lower triangle of a correlation matrix
}
\examples{
get_lower_tri(stats::cor(matrix(rnorm(500), 100, 5)))
}
