% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riingo.r
\name{get_monthly_returns}
\alias{get_monthly_returns}
\title{Get Monthly Returns}
\usage{
get_monthly_returns(symbols, tiingo_token = Sys.getenv("TIINGO_TOKEN"),
  start_date = NULL, end_date = NULL)
}
\arguments{
\item{symbols}{Character vector of symbols}

\item{tiingo_token}{The token for tiigo API}

\item{start_date}{Start date}

\item{end_date}{End date.}
}
\value{
Vector of monthly returns.
}
\description{
Only returns full months.  Dates are last business days of the month.
}
\examples{
get_monthly_returns(symbols="IVV", 
start_date = as.Date("2014-12-31"), 
end_date = as.Date("2018-12-31"))
}
