% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massageData.r
\name{GetFundReturns}
\alias{GetFundReturns}
\title{Get Fund Returns}
\usage{
GetFundReturns(symbols, sdate, edate)
}
\arguments{
\item{symbols}{vector of character strings}

\item{sdate}{start date. Should be the last day of the month before the first month of the return series. 
Example: use 2015-12-31 to generate the first monthly return for January 2016.}

\item{edate}{end date. Last day of the month of the return series.}
}
\value{
3 column tibble with Symbol, Month-End Date, and Ra, the monthly return
}
\description{
Retrieves prices from web and calculates monthly returns.
}
\examples{
GetFundReturns(symbols = c("IVV", "SPY"), sdate=as.Date("2015-12-31"), edate=as.Date("2018-12-31"))
}
