% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDummy.R
\name{makeDummy}
\alias{makeDummy}
\title{makeDummy}
\usage{
makeDummy(tZ)
}
\arguments{
\item{tZ}{An input vector}
}
\value{
Returns dZ, a matrix of size length(tZ)-by-card(tZ) : \describe{
The ij-th element in dZ is 1 if tZ[i] is equal to the j-th largest value of tZ.
And the ij-th element in DZ is 0 otherwise.
The row sum of dZ must be 1 by construction.
}
}
\description{
This function is to make dummy variables using a discrete variable.
}
\examples{
makeDummy(c(1,2,3))

}
\author{
Yujung Hwang, \email{yujungghwang@gmail.com}
}
