% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.r
\name{plot_corr}
\alias{plot_corr}
\title{Plots a correlation matrix}
\usage{
plot_corr(cormat)
}
\arguments{
\item{cormat}{Correlation matrix}
}
\value{
A ggplot object
}
\description{
The upper triangle is a heatmap, the lower triangle is a table.
}
\examples{
plot_corr(cormat)
}
