% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.cov.R
\name{weighted.cov}
\alias{weighted.cov}
\title{weighted.cov}
\usage{
weighted.cov(x, y, w)
}
\arguments{
\item{x}{An input vector to compute a covariance, cov(x,y)}

\item{y}{An input vector to compute a covariance, cov(x,y)}

\item{w}{A weight vector}
}
\value{
Returns an unbiased sample weighted covariance
}
\description{
This function is to compute an unbiased sample weighted covariance. The function uses only pairwise complete observations.
}
\author{
Yujung Hwang, \email{yujungghwang@gmail.com}
}
