% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.var.R
\name{weighted.var}
\alias{weighted.var}
\title{weighted.var}
\usage{
weighted.var(x, w)
}
\arguments{
\item{x}{A vector to compute a variance, var(x)}

\item{w}{A weight vector}
}
\value{
Returns an unbiased sample weighted variance
}
\description{
This function is to compute an unbiased sample weighted variance.
}
\author{
Yujung Hwang, \email{yujungghwang@gmail.com}
}
