% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{factorReturns.Intl}
\alias{factorReturns.Intl}
\title{RAFI International Factor Returns}
\format{A tibble with 381 rows and 15 variables:
\describe{
  \item{date}{date at the end of the month}
  \item{DxUVal}{Int'l Developed Value factor}
  \item{DxULVol}{Int'l Developed Low volatility factor}
  \item{DxUQual}{Int'l Developed Quality factor}
  \item{DxUSize}{Int'l Developed Size factor}
  \item{DxUMom}{Int'l Developed Momentum factor}
}}
\usage{
factorReturns.Intl
}
\description{
A dataset containing the monthly returns of 5 factors across the international developed markets.
}
\keyword{datasets}
