% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{factorReturns.US}
\alias{factorReturns.US}
\title{5 RAFI Factor Returns Across the U.S.}
\format{A tibble with 381 rows and 15 variables:
\describe{
  \item{date}{date at the end of the month}
  \item{USVal}{U.S. Value factor}
  \item{USLVol}{U.S. Low volatility factor}
  \item{USQual}{U.S. Quality factor}
  \item{USSize}{U.S. Size factor}
  \item{USMom}{U.S. Momentum factor}
}}
\source{
RAFI (Research Affiliates)
}
\usage{
factorReturns.US
}
\description{
RAFI U.S. Factor Returns
}
\details{
A dataset containing the monthly returns of 5 factors across the US
}
\keyword{datasets}
